Lint and add pre-commit hooks

+++ contracts/DS/DSNote.sol
@@ -1,13 +1,14 @@
 pragma solidity ^0.5.0;
 
+
 contract DSNote {
     event LogNote(
-        bytes4   indexed  sig,
-        address  indexed  guy,
-        bytes32  indexed  foo,
-        bytes32  indexed  bar,
-        uint              wad,
-        bytes             fax
+        bytes4 indexed sig,
+        address indexed guy,
+        bytes32 indexed foo,
+        bytes32 indexed bar,
+        uint256 wad,
+        bytes fax
     ) anonymous;
 
     modifier note {
