+++ contracts/compound/CompoundProxy.sol
@@ -9,14 +9,14 @@ import "./Exponential.sol";
 import "./StupidExchange.sol";
 import "../constants/ConstantAddresses.sol";
 
+
 /// @title CompoundProxy implements CDP and Compound direct interactions
 contract CompoundProxy is DSMath, Exponential, ConstantAddresses {
-
     /// @notice It will draw Dai from Compound and repay part of the CDP debt
     /// @dev User has to approve DSProxy to pull CDai before calling this
     /// @param _cup Cdp id
     /// @param _amount Amount of Dai that will be taken from Compound and put into CDP
-    function repayCDPDebt(bytes32 _cup, uint _amount) public {
+    function repayCDPDebt(bytes32 _cup, uint256 _amount) public {
         TubInterface tub = TubInterface(TUB_ADDRESS);
         CTokenInterface cDaiContract = CTokenInterface(CDAI_ADDRESS);
 
@@ -26,9 +26,9 @@ contract CompoundProxy is DSMath, Exponential, ConstantAddresses {
         approveTub(WETH_ADDRESS);
 
         // Calculate how many cDai tokens we need to pull for the Dai _amount
-        uint cAmount = getCTokenAmount(_amount, CDAI_ADDRESS);
+        uint256 cAmount = getCTokenAmount(_amount, CDAI_ADDRESS);
 
-        cDaiContract.approve(CDAI_ADDRESS, uint(-1));
+        cDaiContract.approve(CDAI_ADDRESS, uint256(-1));
         cDaiContract.transferFrom(msg.sender, address(this), cAmount);
 
         require(cDaiContract.redeemUnderlying(_amount) == 0, "Reedem Failed");
@@ -37,13 +37,18 @@ contract CompoundProxy is DSMath, Exponential, ConstantAddresses {
         StupidExchange(STUPID_EXCHANGE).getMakerDaiToken(_amount);
 
         // Buy some Mkr to pay stability fee
-        uint mkrAmount = stabilityFeeInMkr(tub, _cup, _amount);
-        uint daiFee = wdiv(mkrAmount, estimatedDaiMkrPrice(_amount));
-        uint amountExchanged = exchangeToken(ERC20(MAKER_DAI_ADDRESS), ERC20(MKR_ADDRESS), daiFee, mkrAmount);
+        uint256 mkrAmount = stabilityFeeInMkr(tub, _cup, _amount);
+        uint256 daiFee = wdiv(mkrAmount, estimatedDaiMkrPrice(_amount));
+        uint256 amountExchanged = exchangeToken(
+            ERC20(MAKER_DAI_ADDRESS),
+            ERC20(MKR_ADDRESS),
+            daiFee,
+            mkrAmount
+        );
 
         _amount = sub(_amount, daiFee);
 
-        uint daiDebt = getDebt(tub, _cup);
+        uint256 daiDebt = getDebt(tub, _cup);
 
         if (_amount > daiDebt) {
             ERC20(MAKER_DAI_ADDRESS).transfer(msg.sender, sub(_amount, daiDebt));
@@ -53,13 +58,18 @@ contract CompoundProxy is DSMath, Exponential, ConstantAddresses {
         tub.wipe(_cup, _amount);
 
         ERC20(MKR_ADDRESS).transfer(msg.sender, ERC20(MKR_ADDRESS).balanceOf(address(this)));
-        ActionLogger(LOGGER_ADDRESS).logEvent('repayCDPDebt', msg.sender, mkrAmount, amountExchanged);
+        ActionLogger(LOGGER_ADDRESS).logEvent(
+            "repayCDPDebt",
+            msg.sender,
+            mkrAmount,
+            amountExchanged
+        );
     }
 
     /// @notice It will draw Dai from CDP and add it to Compound
     /// @param _cup CDP id
     /// @param _amount Amount of Dai drawn from the CDP and put into Compound
-    function cdpToCompound(bytes32 _cup, uint _amount) public {
+    function cdpToCompound(bytes32 _cup, uint256 _amount) public {
         TubInterface tub = TubInterface(TUB_ADDRESS);
         CTokenInterface cDaiContract = CTokenInterface(CDAI_ADDRESS);
 
@@ -72,24 +82,28 @@ contract CompoundProxy is DSMath, Exponential, ConstantAddresses {
         StupidExchange(STUPID_EXCHANGE).getCompoundDaiToken(_amount);
 
         //cDai will try and pull Dai tokens from DSProxy, so approve it
-        ERC20(COMPOUND_DAI_ADDRESS).approve(CDAI_ADDRESS, uint(-1));
+        ERC20(COMPOUND_DAI_ADDRESS).approve(CDAI_ADDRESS, uint256(-1));
 
         require(cDaiContract.mint(_amount) == 0, "Failed Mint");
 
-        uint cDaiMinted = cDaiContract.balanceOf(address(this));
+        uint256 cDaiMinted = cDaiContract.balanceOf(address(this));
 
         // transfer the cDai to the original sender
         ERC20(CDAI_ADDRESS).transfer(msg.sender, cDaiMinted);
 
-        ActionLogger(LOGGER_ADDRESS).logEvent('cdpToCompound', msg.sender, _amount, cDaiMinted);
-
+        ActionLogger(LOGGER_ADDRESS).logEvent("cdpToCompound", msg.sender, _amount, cDaiMinted);
     }
 
     /// @notice Calculates how many cTokens you get for a _tokenAmount
-    function getCTokenAmount(uint _tokenAmount, address _tokeAddress) internal returns(uint cAmount) {
+    function getCTokenAmount(uint256 _tokenAmount, address _tokeAddress)
+        internal
+        returns (uint256 cAmount)
+    {
         MathError error;
-        (error, cAmount) = divScalarByExpTruncate(_tokenAmount,
-             Exp({mantissa: CTokenInterface(_tokeAddress).exchangeRateCurrent()}));
+        (error, cAmount) = divScalarByExpTruncate(
+            _tokenAmount,
+            Exp({mantissa: CTokenInterface(_tokeAddress).exchangeRateCurrent()})
+        );
 
         require(error == MathError.NO_ERROR, "Math error");
     }
@@ -98,43 +112,55 @@ contract CompoundProxy is DSMath, Exponential, ConstantAddresses {
     /// @param _tub Tub interface
     /// @param _cup Id of the CDP
     /// @param _daiRepay Amount of dai we are repaying
-    function stabilityFeeInMkr(TubInterface _tub, bytes32 _cup, uint _daiRepay) public returns (uint) {
+    function stabilityFeeInMkr(TubInterface _tub, bytes32 _cup, uint256 _daiRepay)
+        public
+        returns (uint256)
+    {
         bytes32 mkrPrice;
         bool ok;
 
-        uint feeInDai = rmul(_daiRepay, rdiv(_tub.rap(_cup), _tub.tab(_cup)));
+        uint256 feeInDai = rmul(_daiRepay, rdiv(_tub.rap(_cup), _tub.tab(_cup)));
 
         (mkrPrice, ok) = _tub.pep().peek();
 
-        return wdiv(feeInDai, uint(mkrPrice));
+        return wdiv(feeInDai, uint256(mkrPrice));
     }
 
     /// @notice Returns expected rate for Dai -> Mkr conversion
     /// @param _daiAmount Amount of Dai
-    function estimatedDaiMkrPrice(uint _daiAmount) internal returns (uint expectedRate) {
-        (expectedRate, ) = KyberNetworkProxyInterface(KYBER_INTERFACE).getExpectedRate(ERC20(MAKER_DAI_ADDRESS), ERC20(MKR_ADDRESS), _daiAmount);
+    function estimatedDaiMkrPrice(uint256 _daiAmount) internal returns (uint256 expectedRate) {
+        (expectedRate, ) = KyberNetworkProxyInterface(KYBER_INTERFACE).getExpectedRate(
+            ERC20(MAKER_DAI_ADDRESS),
+            ERC20(MKR_ADDRESS),
+            _daiAmount
+        );
     }
 
     /// @notice Approve a token if it's not already approved
     /// @param _tokenAddress Address of the ERC20 token we want to approve
     function approveTub(address _tokenAddress) internal {
-        if (ERC20(_tokenAddress).allowance(msg.sender, _tokenAddress) < (uint(-1) / 2)) {
-            ERC20(_tokenAddress).approve(TUB_ADDRESS, uint(-1));
+        if (ERC20(_tokenAddress).allowance(msg.sender, _tokenAddress) < (uint256(-1) / 2)) {
+            ERC20(_tokenAddress).approve(TUB_ADDRESS, uint256(-1));
         }
     }
 
     /// @notice Returns current Dai debt of the CDP
     /// @param _tub Tub interface
     /// @param _cup Id of the CDP
-    function getDebt(TubInterface _tub, bytes32 _cup) internal returns (uint debt) {
-        ( , , debt, ) = _tub.cups(_cup);
+    function getDebt(TubInterface _tub, bytes32 _cup) internal returns (uint256 debt) {
+        (, , debt, ) = _tub.cups(_cup);
     }
 
     /// @notice Exhcanged a token on kyber
-    function exchangeToken(ERC20 _sourceToken, ERC20 _destToken, uint _sourceAmount, uint _maxAmount) internal returns (uint destAmount) {
+    function exchangeToken(
+        ERC20 _sourceToken,
+        ERC20 _destToken,
+        uint256 _sourceAmount,
+        uint256 _maxAmount
+    ) internal returns (uint256 destAmount) {
         KyberNetworkProxyInterface _kyberNetworkProxy = KyberNetworkProxyInterface(KYBER_INTERFACE);
 
-        uint minRate;
+        uint256 minRate;
         (, minRate) = _kyberNetworkProxy.getExpectedRate(_sourceToken, _destToken, _sourceAmount);
 
         require(_sourceToken.approve(address(_kyberNetworkProxy), 0));
