+++ contracts/interfaces/SaverExchangeInterface.sol
@@ -1,5 +1,11 @@
 pragma solidity ^0.5.0;
 
+
 contract SaverExchangeInterface {
-    function getBestPrice(uint _amount, address _srcToken, address _destToken, uint _exchangeType) public view returns (address, uint);
+    function getBestPrice(
+        uint256 _amount,
+        address _srcToken,
+        address _destToken,
+        uint256 _exchangeType
+    ) public view returns (address, uint256);
 }
