+++ contracts/mcd/Discount.sol
@@ -1,15 +1,15 @@
 pragma solidity ^0.5.0;
 
+
 contract Discount {
-
     address public owner;
-    mapping (address => CustomServiceFee) public serviceFees;
+    mapping(address => CustomServiceFee) public serviceFees;
 
-    uint constant MAX_SERVICE_FEE = 400;
+    uint256 constant MAX_SERVICE_FEE = 400;
 
     struct CustomServiceFee {
         bool active;
-        uint amount;
+        uint256 amount;
     }
 
     constructor() public {
@@ -20,26 +20,20 @@ contract Discount {
         return serviceFees[_user].active;
     }
 
-    function getCustomServiceFee(address _user) public view returns (uint) {
+    function getCustomServiceFee(address _user) public view returns (uint256) {
         return serviceFees[_user].amount;
     }
 
-    function setServiceFee(address _user, uint _fee) public {
+    function setServiceFee(address _user, uint256 _fee) public {
         require(msg.sender == owner, "Only owner");
         require(_fee >= MAX_SERVICE_FEE || _fee == 0);
 
-        serviceFees[_user] = CustomServiceFee({
-            active: true,
-            amount: _fee
-        });
+        serviceFees[_user] = CustomServiceFee({active: true, amount: _fee});
     }
 
     function disableServiceFee(address _user) public {
         require(msg.sender == owner, "Only owner");
 
-        serviceFees[_user] = CustomServiceFee({
-            active: false,
-            amount: 0
-        });
+        serviceFees[_user] = CustomServiceFee({active: false, amount: 0});
     }
 }
