+++ contracts/scd/MonitorProxy.sol
@@ -6,10 +6,17 @@ import "../DS/DSAuth.sol";
 import "./Monitor.sol";
 import "../constants/ConstantAddresses.sol";
 
+
 /// @title MonitorProxy handles authorization and interaction with the Monitor contract
 contract MonitorProxy is ConstantAddresses {
-
-    function subscribe(bytes32 _cup, uint _minRatio, uint _maxRatio, uint _optimalRatioBoost, uint _optimalRatioRepay, address _monitor) public {
+    function subscribe(
+        bytes32 _cup,
+        uint256 _minRatio,
+        uint256 _maxRatio,
+        uint256 _optimalRatioBoost,
+        uint256 _optimalRatioRepay,
+        address _monitor
+    ) public {
         address currAuthority = address(DSAuth(address(this)).authority());
         DSGuard guard = DSGuard(currAuthority);
 
@@ -20,11 +27,30 @@ contract MonitorProxy is ConstantAddresses {
 
         guard.permit(_monitor, address(this), bytes4(keccak256("execute(address,bytes)")));
 
-        Monitor(_monitor).subscribe(_cup, _minRatio, _maxRatio, _optimalRatioBoost, _optimalRatioRepay);
+        Monitor(_monitor).subscribe(
+            _cup,
+            _minRatio,
+            _maxRatio,
+            _optimalRatioBoost,
+            _optimalRatioRepay
+        );
     }
 
-    function update(bytes32 _cup, uint _minRatio, uint _maxRatio, uint _optimalRatioBoost, uint _optimalRatioRepay, address _monitor) public {
-        Monitor(_monitor).subscribe(_cup, _minRatio, _maxRatio, _optimalRatioBoost, _optimalRatioRepay);
+    function update(
+        bytes32 _cup,
+        uint256 _minRatio,
+        uint256 _maxRatio,
+        uint256 _optimalRatioBoost,
+        uint256 _optimalRatioRepay,
+        address _monitor
+    ) public {
+        Monitor(_monitor).subscribe(
+            _cup,
+            _minRatio,
+            _maxRatio,
+            _optimalRatioBoost,
+            _optimalRatioRepay
+        );
     }
 
     function unsubscribe(bytes32 _cup, address _monitor) public {
